requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Limiter==3.7.0
WTForms==3.1.2
SQLAlchemy==2.0.30
Werkzeug==3.0.3
python-dotenv==1.0.1
pyotp==2.9.0
itsdangerous==2.2.0
email-validator==2.1.1
